+++ packages/core/src/app/checkout/Checkout.tsx
@@ -21,6 +21,7 @@ import { Extension, ExtensionContextProps, withExtension } from '@bigcommerce/ch
 import { ErrorLogger } from '@bigcommerce/checkout/error-handling-utils';
 import { TranslatedString, withLanguage, WithLanguageProps } from '@bigcommerce/checkout/locale';
 import { AddressFormSkeleton, ChecklistSkeleton , LazyContainer, LoadingNotification, LoadingOverlay } from '@bigcommerce/checkout/ui';
+import { navigateToOrderConfirmation } from '@bigcommerce/checkout/utility';
 
 import { withAnalytics } from '../analytics';
 import { StaticBillingAddress } from '../billing';
@@ -49,7 +50,6 @@ import CheckoutStepStatus from './CheckoutStepStatus';
 import CheckoutStepType from './CheckoutStepType';
 import CheckoutSupport from './CheckoutSupport';
 import mapToCheckoutProps from './mapToCheckoutProps';
-import navigateToOrderConfirmation from './navigateToOrderConfirmation';
 
 const Billing = lazy(() =>
     retry(
@@ -180,6 +180,12 @@ class Checkout extends Component<
         this.handleBeforeExit();
     }
 
+    componentDidUpdate(prevProps: WithCheckoutProps): void {
+        if(prevProps.steps.length === 0 && this.props.steps && this.props.steps.length > 0) {
+            this.handleReady();
+        }
+    }
+
     async componentDidMount(): Promise<void> {
         const {
             analyticsTracker,
@@ -270,9 +276,7 @@ class Checkout extends Component<
             });
 
             if (isMultiShippingMode) {
-                this.setState({ isMultiShippingMode }, this.handleReady);
-            } else {
-                this.handleReady();
+                this.setState({ isMultiShippingMode });
             }
 
             window.addEventListener('beforeunload', this.handleBeforeExit);
@@ -330,6 +334,7 @@ class Checkout extends Component<
                 <div className="layout-main">
                     <LoadingNotification isLoading={(!isShowingWalletButtonsOnTop && isPending) || extensionState.isShowingLoadingIndicator} />
 
+                    {/* <Extension region={ExtensionRegion.GlobalWebWorker} /> */}
                     <PromotionBannerList promotions={promotions} />
 
                     {isShowingWalletButtonsOnTop && this.state.buttonConfigs?.length > 0 && (