+++ packages/core/src/app/common/utility/index.ts
@@ -7,6 +7,5 @@ export { default as isRecordContainingKey } from './isRecordContainingKey';
 export { default as isMobile } from './isMobile';
 export { default as isApplePayWindow } from './is-apple-pay-window';
 export { default as shouldFilterApplePay } from './should-filter-apple-pay';
-export { default as isBuyNowCart } from './isBuyNowCart';
 export { default as isFloatingLabelEnabled } from './isFloatingLabelEnabled';
 export { default as isExperimentEnabled } from './isExperimentEnabled';