+++ packages/core/src/app/order/OrderConfirmationSection.tsx
@@ -1,6 +1,6 @@
 import React, { FunctionComponent } from 'react';
 
-const OrderConfirmationSection: FunctionComponent = ({ children }) => (
+const OrderConfirmationSection: FunctionComponent<{ children?: React.ReactNode }> = ({ children }) => (
     <section className="orderConfirmation-section">{children}</section>
 );