+++ packages/utility/src/index.ts
@@ -1,0 +1,2 @@
+export { default as isBuyNowCart } from './isBuyNowCart';
+export { default as navigateToOrderConfirmation } from './navigateToOrderConfirmation';